Enabling failing of tests if circular references are found (Bug 4766)
A NOX user noticed that NOX is creating some circular references that result
in a bunch of warnings at the end of the program (see Bug 4766).  These
warnings show memory leaks that are occuring due to circular RCP objects.

I then added a grep to the CTest tests looking for this warning message and I
found a bunch of errors coming from packages including NOX, Stratimikos,
Ifpack, ML, CTrilinos, MOOCHO, and Rythmos (and that is just Primary Stable
code).  We already know that Sundance has a bunch of serious memory leaks of
this nature.  Some of these failures can be seen at:

    http://trilinos.sandia.gov/cdash/index.php?project=Trilinos&date=2010-04-29&display=project&filtercount=4&showfilters=1&filtercombine=and&field1=buildname/string&compare1=61&value1=CircularReferenceFailures&field2=site/string&compare2=61&value2=godel.sandia.gov&field3=buildstarttime/date&compare3=83&value3=2010-04-29T17:53:47%20MDT&field4=buildstarttime/date&compare4=84&value4=2010-04-29T20:08:51%20MDT&collapse=0

(NOTE: The failures for Zoltan are not related).

What this commit does is it turns on a CTest grep that looks for the warnings
and if it finds it, it fails the test.  This is now the default behavior.
However, this can be turned off by passing in
DISABLE_CIRCULAR_REF_DETECTION_FAILURE to the PACKAGE(...) macro in the
top-level package CMakeLists.txt file.

Note that the top-level option
Trilinos_ENABLE_CIRCULAR_REF_DETECTION_FAILURE=ON can also be set to turn on
this checking for all of Trilinos.