Changes for 64 bit global IDs and global numbers and counts.
Many changes and fixes.  Some are just to silence compiler warnings.
But most implement ZOLTAN_GNO_TYPE (a integer data type that is
32 bits on a 32 bit machines and 64 bits on a 64 bit machine) and
ZOLTAN_ID_TYPE (which is set when zoltan is compiled and defaults
to unsigned int).

I also added functions in third_library.c to print out the graph
and matrix structures for debugging.