Changed this test to compute better eigenvalues.
I changed this test to use the Cayley transformation to compute
eigenvalues so that the bifurcation is exhibited.  I also changed the
number of eigenvalues to be odd to catch problems in Anasazi for
splitting complex-conjugate pairs of eigenvalues.  To get the correct
eigenvalues, I also had to change the sign of the mass matrix.