Added null vector update and mass matrix scaling options.
I added options to update the null vector after each continuation step,
and also scale it by the range of the mass matrix, similar to the
minimally augmented method.  I also fixed what appeared to be two small
copying bugs in the minimally augmented constraint with regards to these
options.