Improves usage of RequestHandler interface for getting additional parameters
This change replaces the use of "getRequestedParameters" and "updateRequestedParameters"
in many places.  The classes requiring this functionality now use the RequestHandler
listener interface. This streamlines, in a considerable way, the implementation of
the preconditioners.  Also removes a lot of ParameterList code, which cleans things
up a little.

Also added the StaticRequestCallback class. This is useful for
testing the request callback interface. It may also be useful early
in the development process.