Updating testing on the mac machines
Disabled the serial debug dev build on s903186 and disabled the mpi opt dev
build on s909348. These machines were running the same set of tests for
comparison reasons. Now that they are running the same set of tests with one
exception I split them up so as not to overload anyone machine. Both machines
were running late into the morning. The exception in the tests that are run is
that s909348 doesn't have cpp unit on it, however those tests will still be run
on s903186 and cppunit will be installed on s909348 in the near future.