Fix the "no Update.xml" issues on TrilinosDriver dashboards.
git_exe was not being found correctly because of the local
variable set to the empty string used in the find_program call.
Set it explicitly to git_exe-NOTFOUND prior to find_program.

Also: remove the TODO comment and do the TODO item... :-)
Added the ability to override the "model" used for the driver
dashboard by adding the variable TDD_CTEST_TEST_TYPE. Set
it in the environment of the calling machine to Nightly to
run a nightly dashboard. Defaults to Experimental.