Fix configure issues with Trios
Added netcdf to the nightly builds for s909348 since Trios requires this tpl.

Also fixed an issue in ioss where an optional tpl wasn't set up and therefore a
required variable didn't exist. The tpl xdmf has been added, but mostly in as a
stub capability. This is due to the fact that it relies on a version of hdf5
that isn't supported yet and expat. For this reason it was made TS. Once we get
those tpls on a machine and testing we can move this to SS, however, epetraext
will need to be updated to support the newer hdf5. In the mean time having this
tpl here fixes the configure error for ioss without removing the capability.