Fix an issue with tarball size doubling
The change to the package controlled exclude list had an issue where the exclude
list wasn't fully created before the point at which the cpack files were
created. This was because the tarball set up was done before packages were
looped over. Since the files that each package excludes is now done inside each
package this meant that their exclude files weren't being picked up. This
slipped through because the testing for exclude change was done based on the
list of files to exclude after the packages had been looped over. Moving the
tarball setup step to be after when each package is handled fixes this. This has
been tested by comparing the tarball generated for the master branch and the
tarball generated for 10.4.1 and while the tarball has increased in size it can
all be attributed to new things added to Trilinos since 10.4.

I also fixed an issue where the source directory that a tarball was generated
from had to be named "Trilinos". This caveat bit me so I did some research and
found that there is a variable that points to the top level of the source dir
regardless of where you are currently in the tree. It is now safe to build a
source package from any source directory without fear of having excludes not
being obeyed.