Adding topology aware partitioning to Zoltan's HIERARCHICAL method
Some of our machines are hard coded into the platform_specs array
defined at the top of hier.c.  These can be specified at run
time with the PLATFORM_NAME parameter.

Other machines can be specified at run time with the TOPOLOGY
parameter.  This is a string giving the number of siblings within
a node or a machine at each level.

So, a quad-socket dual-core machine would be "4,2".
A quad-core workstation would be "4".

Feel free to add other machines t the platform_specs.

This seems to work but I'm checking it in so I can do more testing.

See example/C/simpleHIER.c for an example of using Zoltan-assisted
hierarchical partitioning.