Allow the cron driver script to be run from any where.
There was an issue where the generic cron driver script expected to be run out
of a repository that was named "Trilinos". This actually was only an issue if
you needed to update cmake though so it wasn't always being tripped. It is now
possible to run all stages of the driver from any repository.