Fixing behavior of checkin-test.py script for changes to 'ifpack2' matching 'Ifpack' instead of 'Ifpack2' (bug 4931)
In bug 4913, it is described how the checkin-test.py script would look at:

   packages/ifpack2/CMakeLists.txt

and report that this was 'Ifpack' modified.  The problem is that the logic
that set up would match 'packages/ifpack' in the packages list before it
mached 'packages/ifpack2' because Ifpack can before Ifpack2 in the packages
list and the logic I had would do an

  re.match('packages/ifpack', 'packages/ifpack2/CMakeLists.txt')

which matches and returned 'Ifpack' is the modified package.

I added a few unit tests to demonstrate the wrong behavior and then I fixed
the code so that 'packages/ifpack2/CMakeLists.txt' would only match the full
directory name 'packages/ifpack2' by essentially doing:

  re.match('packages/ifpack/', 'packages/ifpack2/CMakeLists.txt')

which would not but doing:

  re.match('packages/ifpack2/', 'packages/ifpack2/CMakeLists.txt')

does match.

All of the unit tests pass.

Build/Test Cases Summary
Enabled Packages: TrilinosFramework
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=81,notpassed=0 (0.96 min)
1) SERIAL_RELEASE => passed: passed=81,notpassed=0 (0.90 min)