Disable a framework test when in release mode
The test TrilinosPackageFilePathUtils_UnitTests relies on a package FEApp being
configured, however in a release tarball FEApp doesn't exist since it currently
doesn't have copyright. The test in question uses FEApp as an interesting corner
case that I feel should be tested so just removing the dependence on FEApp
doesn't make sense. Disabling this test for release mode is the only other clean
way to ensure our testing doesn't cause known errors on user machines if they
choose to run the tests. This does mean that this test won't be run on a release
branch at all, however, there should be so little development on the cmake
framework that the risk of a regression here is low.