Restore a PyTrilinos epetra test that had been temporarily disabled...
The test had been tripping on an epetra behavior change that was difficult
to diagnose and correct. This test was disabled during the branching process,
until it could be determined whether the epetra behavior was correct, or
the test's expectations were correct. Epetra's behavior (CrsGraph indices bug)
has now been corrected.