(From Sierra) This is a series of commits to switch SIERRA from Zoltan stand alone version 2.01 to Trilinos/Zoltan 3.4.
There are quite a few changes to delete the old source files from TPLs/zoltan and add in the new
source files in Trilinos/packages/zoltan.  At the same time there are a few minor coding changes
needed to use the newer version of Zoltan.

NOTE: Deleting the source files will not delete object and library files associated with old Zoltan 2.01.
Care must be taken to remove the old libraries to avoid linking two versions of Zoltan.  If using the
bake command the --super-clean option should do it.

(A) Bjam changes to enable use of Trilinos/Zoltan in SIERRA

Change Jam files to not compile the old Zoltan library
and instead compile the new Zoltan library that is in Trilinos.
This effects all of the products that use Zoltan.

(B) Delete stand alone Zoltan to enable use of Trilinos/Zoltan

Remove the 2.01 Zoltan source so there is no chance of compiling it or
using the headers.

(C) Add links from flat headers to Zoltan to enable use of Trilinos/Zoltan

There are quite a few Zoltan headers which are normally scattered through
the Zoltan source tree.  But for Trilinos these are all soft linked into a single
flat directory.  It seems that none of the file names conflict, so this should work.

(D) Code changes from Zoltan version 2.01 to 3.4 to enable use of Trilinos/Zoltan

There was one call, Zoltan_LB_Eval, that changed.  Before it returned a long list
of values from the evaluation of various mesh quality measurements.  In Zoltan 3.4
a structure is returned instead.

(E) Sierra changes to enable use of Trilinos/Zoltan

There are still a couple of files that the Sierra tools use that need to be fixed
up, like package_definitions.py.