(From Sierra) STK_IO: Sideset fixes
* Fix handling of "flat sidesets" which is when the sideset
faces/edges are defined in a single part with no subsets instead of a
sideset with multiple topologies which could be in a part that has a
subset for each different face/edge topology.

* Fix the determination of the parent element for an
edge/face. Existing code could miss a valid parent element if the
elements were processed in an incorrect order.  It was also incorrect
for the case where there was a single element touching the face/edge.