Changed an iterator to a const_iterator to try to fix IBM build.
Their was an implicit cast that was potentially causing a problem on IBM
machines.

Comment from Ross Bartlett: I can not log onto this IBM machine yet so this
has not been tested on the IBM yet.  However, it is harmless to go ahead and
try to push this and see if this fixes the problem on the machine.

Build/Test Cases Summary
Enabled Packages: Teuchos
0) MPI_DEBUG => passed: passed=70,notpassed=0 (1.95 min)
1) SERIAL_RELEASE => passed: passed=67,notpassed=0 (1.13 min)
Other local commits for this build/test group: b6bba2d