Adding ability to tentatively enable a TPL, adding MUST_FIND_ALL_LIBS option (bug 5039)
Here, I have enhanced the TPL system to allow TPLs to be automatically
tentatively enabled.  This allows some TPLs to be automatically attempted to
be enabled but if everything is not found, then it will automatically disable
the TPL.

I also added the MUST_FIND_ALL_LIBS option to require that at least one of the
libraries in the names list be found or it is an error.  This should most
likely be the default behavior but I did not want to break backward
compatibility so I made this an extra option.  This issue became apparent when
I was experimenting with creating the BinUtils TPL.