Adding if statement around LIST(REMOVE_DUPLICATES ...) in case of an empty list (bug 5039)
Eric Phipps reported that when a tentativly enabled TPL is being processed
that if it fails to find any header files then it reports the fatal error:

CMake Error at cmake/package_arch/TPLDeclareLibraries.cmake:284 (LIST):
  list sub-command REMOVE_DUPLICATES requires list to be present.
Call Stack (most recent call first):
  cmake/TPLs/FindTPLBinUtils.cmake:3 (TPL_DECLARE_LIBRARIES)
  cmake/package_arch/PackageArchGlobalMacros.cmake:693 (INCLUDE)
  CMakeLists.txt:179 (PACKAGE_ARCH_PROCESS_ENABLED_TPLS)

The issue is that for a explicilty enabled TPL, the configure process would
have stopped when a header was not found.  However, with a tentatively enabled
TPL this LIST(REMOVE_DUPLICATES ...) statement may execute with an empty list
which is currently an error with CMake.

Here I added a if test around this that should fix the problem.

Build/Test Cases Summary
Enabled Packages: Teuchos
0) MPI_DEBUG => passed: passed=70,notpassed=0 (3.23 min)
1) SERIAL_RELEASE => passed: passed=67,notpassed=0 (1.19 min)
Other local commits for this build/test group: f1c5e2d