(From Sierra) Add use case 8 to stk rebalance: forgot to commit all the changes...
Use Case 8 is rebalancing based on a sideset.

Also changed the weight field parameter to be a
pointer for the reblance_needed function to optionally
pass NULL just as it already was for the full_rebalance
function.  Also changed the Parameter object inside
rebalance from a reference to a pointer since this
is a class with lots of virtual functions that should
be called through a pointer to take advantage of
the virtualness.