Added Velocity Verlet time integrator as a Piro Solver.
Base on conversations with Mike Parks, we decided to
prototype integrators for second order systems in Piro.
No sensitivity analysis or other complications yet,
and no good handling of the velocity unknown.

Greated MockModelEval_B in the test directory that
for now returns f=-1. This is interpreted as the
accelration, and the new test problem verifies
the solution  x = t - t*t/2.