Fixed an OpenMP pragma scoping error, added EPETRA_HAVE_OMP in addition to Epetra_HAVE_OMP.
The Intel compiler complained about the scoping of one of the shared variables.  This is fixed (in Epetra_CrsGraph.cpp)
I added EPETRA_HAVE_OMP as a definition for when Epetra has OpenMP support enabled.  All caps
is consistent with other defines.  I did not remove the Epetra_HAVE_OMP, just in case someone else is already
using it.  Both will always have the same state.