Improved performance of iterating through Cijk tensor.
Recently the data structures for the sparse Cijk tensor were changed to
a set of nested maps.  This allowed for very flexible insertion of
entries and easy search, but resulted in slow iteration through the
tensor (this is critical for computing SG projections of operations such
as *).  Here I have added an additional set of data structures (similar
to the original) and corresponding iterators that allow for much faster
transveral.  There is now a fillComplete() method that must be called
once all of the entries are added, which copies the data into the new
data structures, and must be called before any find() methods are called
or iterators are used.  Checks have been put in place to make sure this
happens (for debug builds).  Once fillComplete() is called, the flexible
insertation data structures are cleared.  The nice thing about using
iterators is I was able to completely change these data structures
without changing any code that uses the tensor.