Removed unnecessary cmakedefine's that have no corresponding test.
This fixes a bunch of unnecessary recompiliation that happens when
turning on Dakota support, since it does have tests for some of these
defines.  These appear to be holdovers from the autotools build system.
There are still a lot more of these, I just removed the ones that were
causing me problems.