Installation testing fixes for Thyra
A couple tests in Thyra were not picking up the correct include directories in
installation testing mode. The needed directories were added for the tests.

There was also a header file that was including a cpp file. The cpp file was not
being installed and so the header wasn't able to compile. I believe that the cpp
file was not needed as its corresponding header file was being included and
there was no impact to the build after I removed it.