changed pamgen include file to use the include path instead of being hard coded
the pamgen test driver was hard coding the location of pamgen_config.h file to
the binary tree of a build. This was causing the installation testing to fail
when looking for it since the file was moved to be in the same directory as all
other include files. I removed the path information to the include so that the
paths given by include directories will be used to find the file.