(From Sierra) STK_MESH: Replace DefaultFEM with FEMMetaData
In order to get rid of DefaultFEM, the stk_io classes need to
be able to take a fem::FEMMetaData instead of just a mesh::MetaData.
The problem is that it is hard to change all the code throughout all
of the stk classes at once time.  So first duplicate all of the
stk_io functions to take either a mesh::MetaData or a fem::FEMMetaData
and then once everything that calls it is converted over, remove
the mesh::MetaData option.  This commit punts on most of the hard
stuff and just gets a mesh::MetaData off of the fem::FEMMetaData
and calls the old functions. Better code will be written once the
mesh::MetaData is removed.