o unconditionally clear culled flag before optimization o move setting of other flags from PatchData to Instruction o fix bug doing non-global smooth with Settings::FixedVertexMode != Settings::FIXED_FLAG o fix bug where stale values in PatchData::byteArray get re-used if PatchData::fill is called more than once
Historically Mesquite has expected the application to provide the "vertex byte"
already zero'd when optimization begins and if Mesquite does culling, in which
case it is setting a bit in that value, it re-zeros the value for each vertex
after the optimization. This is somewhat fragile because a) the application will
see strange behavior if they do not zero the value initially and b) mesquite
will not reset the value if an optimization fails.  While it is slightly more
expensive, it is a whole lot safer to clear that bit at the start of each
optimization.  I had previously added code to clear the culled bit on each
vertex iff the culling is being used.  However, that was incomplete because
other code (e.g. VertexPatches) will always check the flag because it doesn't
know whether or not culling is being used.  Now we always clear the culled bit
at the start of every optimization rather than resetting it after.

Also, as long as the code is messing with the vertex byte at the beginning of
each optimization, copy the values from Mesh::vertices_get_fixed_flag and
Mesh::vertices_get_slaved_flag (or the calculated values if the Settings
indicate that) into the vertex byte at the same time.  This might make things
a little bit faster when doing a non-global optimization because the PatchData
doesn't need to gather information from all of those locations when populating
its contents.  However, the main benefit of this is that it consolidates the
logic for when a vertex is considered fixed or slaved, which exposed bugs in a
few other places that assumed this was always determined by the
Mesh::vertices_get_*_flag() values.

The above change also required that the MSQ_PATCH_VTX flag be changed to
MSQ_PATCH_FIXED, having the opposite meaning.  Now only vertices designated
as fixed by the application have MSQ_HARD_FIXED set.  Vertices that are fixed
only because they are on the boundary of the patch are marked with
MSQ_PATCH_FIXED.  A vertex is considered fixed if either flag is set.  This
allows PatchData to write back the vertex byte without worrying about whether
or not the MSQ_HARD_FIXED flag was set because the application told us the
vertex is fixed or when the patch was populated.  It is now only set in the
former case with the latter indicated by MSQ_PATCH_FIXED.