Add an installtion testing build to trilinos-test2
This is a first pass at installation testing.

Added a nightly test that will create a tarball from the master branch, then
from that build an installation tree which will be used by a second test to
build against.

The tarball and installation build are currently handled by a python script that
was originally designed to work as a stand alone script to create a tarball and
installation from it. I would like to revisit this at some point and make it
more compatible with the tdd system so that we can get detailed information
about the individual builds including a normal build/test cycle run compeletely
from the tarball.