Disable Pamgen tests for AIX
There were a couple tests that are failing on the AIX build that were disabled
to make the AIX build will be clean.

The following tests were disabled:
Pamgen_run_test_script_16
Pamgen_run_test_script_18

Disabling these tests was a little bit more complicated that the other packages.
Pamgen uses a wrapper for its tests since it has a single driver that it uses
and changes the test parameters by passining in command line arguements to the
driver. The wrapper didn't allow passing in options such as the required
XHOSTTYPE so I had to add that feature to its wrapper. I added all the HOST
related options as I expect we may need to make this kind of change for other
machines in the future. As a side effect of these changes the arguments that
are passed in now need to have "ARGS" put in before them. This is because ARGN
was being used before and it can now contain things that should not be passed to
the drivers. I have updated all existing tests to work with this modified
wrapper so it should be a simple matter to base any new tests off of the
examples that are already there.

There were two such wrappers, one for the int driver and one for the long long
driver. to prevent duplication of these changes in both those wrappers I added
another option called "LONG_LONG" that will create a test for the long long
driver instead of the int driver.