updates to installation testing to support backwards compatibility testing
These changes set the ground work for a backwards compatibility nighlty build.
The backward compatibility build relies on the installation testing work to
build against an install of a newer version of Trilinos. This uses the same
install of Trilinos that the nightly installation testing uses so that had to be
moved to a more general location. The actual backwards compatibility test will
build a previous version of trilinos against the install. In this case it is set
up to build 10.8 against development. In the future as we add new branches that
support installation testing we will add them as well.

Since the installation testing work only exists in the master branch right now
there is nothing we can test for backwards compatibility, but once 10.8(which
will be the first release with installation testing) is released it should only
take enabling the commented out code in the
cmake/ctest/drivers/trilinos-test2/CMakeLists.txt file to get it working.

this capability was tested by making a private copy of the Trilinos repository
and then adding a few branches to it that were based off of master so that they
would have the installation testing feature. Then from a development branch
tests were run that installed the development branch, then ran tests from these
mock branches. This showed that the tests could get a different branch to
successfully link against the install, but clearly couldn't show any potential
issues with backwards compatibility as the branch it was using was at basically
the same commit as master.