Added the name(s) methods to the base class for DataLayout.
I was going to use multiple inheritance (mix in interfaces) to support the name(s) function on data layout to keep the base class clean.  However it didn't seem that adding in the complexity and required dynamic cast was worth the effort.  In the future, if we find that more methods must be added to data layout, then I will revisit and maybe move into a separate class.